tests/wpt/web-platform-tests/speculation-rules/prerender/prefetch.https.html

<!DOCTYPE html>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/speculation-rules/prerender/resources/utils.js"></script>
<body>
<script>
setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  const {tryToActivate, getNetworkRequestCount} =
      await create_prerendered_page(t, {}, {prefetch: true}, {});

  assert_equals(await tryToActivate(), 'activated');
  assert_equals(await getNetworkRequestCount(), '2', 'Network request count');
}, "Prerender navigation requests don't use prefetched results (for now)");
</script>
</body>
